feat(gameplay): close spec 5.1 armor pip rail + split record sheet panels
- Add ArmorPipRail sub-component rendering per-location armor and
  internal structure as compact 8x8 dots, reusing the ArmorPip visual
  palette (green/amber/red/gray) at a size appropriate for the dense
  combat action panel. Satisfies the delta spec requirement that the
  action panel "include armor pips for all eight locations".
- LocationStatusRow now renders armor + optional rear-armor + structure
  pip rails below the existing numeric row.
- Extract WeaponRow + InlineAmmoCounter + WeaponStatusBadges out of
  RecordSheetPanels.tsx into WeaponRow.tsx so each file stays under
  the 400-line lint cap.
- Graduate eight architectural decisions discovered during execution
  into design.md (action-panel-is-record-sheet, lg drawer breakpoint,
  ArmorPipRail rationale, SimpleHeatDisplay vs HeatTracker deferral,
  InlineAmmoCounter rationale, unitSpas projection, phase-banner side
  tint, event log phase+actor columns).
- Merge the five ADDED requirements from the delta spec into the
  tactical-map-interface baseline: Token Selection Binding, Action
  Panel Contract, Phase and Turn HUD, Event Log Panel, Unit Token
  Facing Indicator Binding.
- Reword task 6.1 with a deferral note pointing readers to the design
  decision explaining why SimpleHeatDisplay stays instead of swapping
  to HeatTracker (which is construction-tool-only UX).

Hook-equivalent checks run manually (worktree cannot spawn the husky
pre-commit directly): oxlint --fix + oxfmt + tsc --noEmit + npm run
build all pass on the staged set. Gameplay jest suite 301/301 pass.
openspec validate --strict passes.
